[nextion] Extract common `upload_end_` function to shared file (#9155)
